requirements.txt
pydicom
numpy
matplotlib
pyyaml